% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Visualize a pipeline}
\usage{
draw(pipeline, ...)
}
\arguments{
\item{pipeline}{an instantiated \code{pipeline} object}

\item{...}{reserved for future use}
}
\description{
This function draws a DAG of the existing pipeline flow. For additional information see \code{igraph::spec_viz}
}
