% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{execute}
\alias{execute}
\title{execute a pipeline component}
\usage{
execute(x, envir)
}
\arguments{
\item{x}{component node to be executed}

\item{envir}{calling environment}
}
\description{
execute a pipeline component
}
