% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{Joiner}
\alias{Joiner}
\title{Initialize a Joiner object}
\usage{
Joiner(method)
}
\arguments{
\item{method}{function to join incoming objects together}
}
\value{
a \code{sewage_joiner} object
}
\description{
The \code{Joiner} takes in objects and joins them according to a
defined method into a single node.
}
\note{
additional arguments to be passed to \code{method} should be passed in the
\code{...} of [add_node()]
}
\examples{
pipeline = Pipeline() |>
    add_node(Joiner(method = rbind), name = "Joiner", input = c("file1", "file2"))
}
