% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{print.sewage_pipeline}
\alias{print.sewage_pipeline}
\title{Printing Pipelines}
\usage{
\method{print}{sewage_pipeline}(x, ...)
}
\arguments{
\item{x}{a [Pipeline()] object}

\item{...}{not used}
}
\value{
formatted sewage pipeline output
}
\description{
print a sewage pipeline

this will print all nodes and theis inputs in the pipeline.
   Once the pipeline has been executed, print will show the outputs available
   through [pull_output()]
}
\examples{
pipeline = Pipeline() |>
    add_node(component = head, name = "Head", input = "file")
print(pipeline)
}
