% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pull_output}
\alias{pull_output}
\alias{pull_output.sewage_pipeline}
\title{Extract output components from a pipeline}
\usage{
pull_output(x, component, ...)

\method{pull_output}{sewage_pipeline}(x, component, ...)
}
\arguments{
\item{x}{an executed pipeline object}

\item{component}{a character string specifying which output component to pull}

\item{...}{reserved for future use}
}
\value{
output from a terminating node of an executed sewage pipeline
}
\description{
Extract output components from a pipeline
}
\examples{
pipeline = Pipeline() |>
    add_node(component = head, name = "Head", input = 'file')
result = run(pipeline, file = iris)
pull_output(result, "Head")
}
