% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run}
\alias{run}
\title{Run a pipeline}
\usage{
run(pipeline, ...)
}
\arguments{
\item{pipeline}{an initialized pipeline object}

\item{...}{parameter(s) to pass to first node of the pipeline. This should match the `input` parameter of `add_node` of the first node.
In the case that you have multiple inputs, each argument should match the name of a starting node in your pipeline.}
}
\description{
This function is the extry point for executing a pipeline object
}
