% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pull_output}
\alias{pull_output}
\alias{pull_output.sewage_pipeline}
\title{Extract output components from a pipeline}
\usage{
pull_output(x, component, ...)

\method{pull_output}{sewage_pipeline}(x, component, ...)
}
\arguments{
\item{x}{an executed pipeline object}

\item{component}{a character string specifying which output component to pull
#' \dontrun{
pipeline = Pipeline()
pipeline = pipeline |>
    add_node(name = 'Splitter', component = Splitter(), input = 'file')
result = run(pipeline, file = mtcars)
result$outputs$Splitter.output_1
result$outputs$Splitter.output_2
}}
}
\description{
Extract output components from a pipeline
}
