% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pull_output}
\alias{pull_output}
\alias{pull_output.sewage_pipeline}
\title{Extract output components from a pipeline}
\usage{
pull_output(x, component, ...)

\method{pull_output}{sewage_pipeline}(x, component, ...)
}
\arguments{
\item{x}{an executed pipeline object}

\item{component}{a character string specifying which output component to pull}

\item{...}{reserved for future use}
}
\description{
Extract output components from a pipeline
}
